GerhardBot1September
A chatbot that is specifically designed to help people that are suffering with depression, deal with their individual problems. GerhardBot1 can also help shy teens deal with their problems by listening to the individual's needs and then finding a solution to a problem or set of circumstances. It uses an extensive AIML code database to search and research specific problems before coming back with an answer. This takes on average 2.5 seconds.